+++ R/Ice.R
@@ -20,11 +20,12 @@
 #' 
 #' For Ice$new():
 #' \describe{
-#' \item{predictor}{Object of type \code{Predictor}. See \link{Predictor}.}
-#' \item{feature}{The feature name or index for which to compute the individual conditional expectations.}
-#' \item{grid.size}{The size of the grid for evaluating the predictions}
-#' \item{center.at}{The value for the centering of the plot. Numeric for numeric features, and the level name for factors.}
-#' \item{run}{logical. Should the Interpretation method be run?}
+#' \item{predictor: }{(Predictor)\cr 
+#' The object (created with Predictor$new()) holding the machine learning model and the data.}
+#' \item{feature: }{(`character(1)`)\cr The feature name or index for which to compute the individual conditional expectations.}
+#' \item{grid.size: }{(`numeric(1)`)\cr The size of the grid for evaluating the predictions}
+#' \item{center.at: }{(`numeric(1)`)\cr The value for the centering of the plot. Numeric for numeric features, and the level name for factors.}
+#' \item{run: }{(`logical(1)`)\cr Should the Interpretation method be run?}
 #' }
 #' 
 #' 
@@ -124,81 +125,13 @@ NULL
 #' @export
 
 Ice = R6::R6Class("Ice",
-  inherit = PartialDependence,
+  inherit = Partial,
   public = list( 
     initialize = function(predictor, feature, grid.size = 20, center.at = NULL, run = TRUE) {
-      checkmate::assert_number(center.at, null.ok = TRUE)
-      private$anchor.value = center.at
-      super$initialize(predictor = predictor,  feature=feature, run = run, 
-        grid.size = grid.size)
-    }, 
-    center = function(center.at) {
-      private$anchor.value = center.at
-      private$flush()
-      self$run()
-    }
-  ),
-  private = list(
-    generatePlot = function() {
-      p = ggplot(self$results, mapping = aes_string(x = self$feature.name, 
-        y = "y.hat", group = "..individual"))
-      if (self$feature.type == "numerical") p = p + geom_line(alpha = 0.5)
-      else if (self$feature.type == "categorical") {
-        p = p + geom_line(alpha = 0.2) 
-      }
-      if (private$multiClass) {
-        p = p + facet_wrap("..class.name")
-      } 
-      p +  scale_y_continuous(expression(hat(y)))
-    }, 
-    intervene = function() {
-      dataDesign = super$intervene()
-      if (!is.null(private$anchor.value)) {
-        dataDesign.anchor = private$dataSample
-        dataDesign.anchor[, self$feature.index] = private$anchor.value
-        private$dataDesign.ids = c(private$dataDesign.ids, 1:nrow(private$dataSample))
-        dataDesign = rbind(dataDesign, dataDesign.anchor)
-      }
-      dataDesign
-    },
-    aggregate = function() {
-      X.id = private$dataDesign.ids
-      X.results = private$dataDesign[, self$feature.index, with = FALSE]
-      X.results$..individual = X.id
-      if (private$multiClass) {
-        y.hat.names = colnames(private$qResults)
-        X.results = cbind(X.results, private$qResults)
-        X.results = melt(X.results, variable.name = "..class.name", 
-          value.name = "y.hat", measure.vars = y.hat.names)
-      } else {
-        X.results$y.hat= private$qResults
-        X.results$..class.name = 1
-      }
-      
-      if (!is.null(private$anchor.value)) {
-        anchor.index = which(X.results[,self$feature.name, with=FALSE] == private$anchor.value)
-        X.aggregated.anchor = X.results[anchor.index, c("y.hat", "..individual", "..class.name"), with = FALSE]
-        names(X.aggregated.anchor) = c("anchor.yhat", "..individual", "..class.name")
-        # In case that the anchor value was also part of grid
-        X.aggregated.anchor = unique(X.aggregated.anchor)
-        X.results = merge(X.results, X.aggregated.anchor, by = c("..individual", "..class.name"))
-        X.results$y.hat = X.results$y.hat - X.results$anchor.yhat
-        X.results$anchor.yhat = NULL
-      }
-      
-      # Remove class name column again if single output
-      if (!private$multiClass) {
-        X.results$..class.name = NULL
-      }
-      
-      X.results
-    },
-    anchor.value = NULL
-  ),
-  active = list(
-    center.at = function(x) {
-      if(!missing(x)) warning("Please use $center() to change the value.")
-      return(private$anchor.value)
+      .Deprecated("Partial", msg = "The use of the 'Ice' class is deprecated and it will 
+        be removed starting from version 0.4. Please use the 'Partial' class instead.")
+      super$initialize(predictor, feature, ice = TRUE, aggregation = "none", 
+        grid.size = grid.size, center.at = center.at, run = TRUE)
     }
   )
 )
@@ -208,6 +141,7 @@ Ice = R6::R6Class("Ice",
 #' plot.Ice() plots the individiual expectation results from an Ice object.
 #' 
 #' @param x An Ice R6 object
+#' @param rug [logical] Should a rug be plotted to indicate the feature distribution?
 #' @return ggplot2 plot object
 #' @seealso 
 #' \link{Ice}
@@ -224,7 +158,7 @@ Ice = R6::R6Class("Ice",
 #' # Plot the results directly
 #' plot(ice)
 #' }
-plot.Ice = function(x) {
-  x$plot()
+plot.Ice = function(x, rug = TRUE) {
+  x$plot(rug)
 }