+++ R/PartialDependence.R
@@ -19,7 +19,8 @@
 #' 
 #' For PartialDependence$new():
 #' \describe{
-#' \item{predictor}{Object of type \code{Predictor}. See \link{Predictor}.}
+#' \item{predictor: }{(Predictor)\cr 
+#' The object (created with Predictor$new()) holding the machine learning model and the data.}
 #' \item{feature}{The feature name or index for which to compute the partial dependencies. 
 #' Either a single number or vector of two numbers.}
 #' \item{grid.size}{The size of the grid for evaluating the predictions}
@@ -117,137 +118,14 @@ NULL
 
 #' @export
 
-PartialDependence = R6::R6Class("PartialDependence", 
-  inherit = InterpretationMethod,
-  public = list(
-    grid.size = NULL, 
-    feature.index = NULL, 
-    feature.name = NULL,
-    n.features = NULL, 
-    feature.type= NULL,
+PartialDependence = R6::R6Class("PartialDependence",
+  inherit = Partial,
+  public = list( 
     initialize = function(predictor, feature, grid.size = 20, run = TRUE) {
-      feature = private$sanitizeFeature(feature, predictor$data$feature.names)
-      checkmate::assert_numeric(feature, lower = 1, upper = predictor$data$n.features, min.len = 1, max.len = 2)
-      checkmate::assert_numeric(grid.size, min.len = 1, max.len = length(feature))
-      if (length(feature) == 2) checkmate::assert_false(feature[1] == feature[2])
-      super$initialize(predictor)
-      private$setFeatureFromIndex(feature)
-      private$set.grid.size(grid.size)
-      private$grid.size.original = grid.size
-      if(run) self$run()
-    }, 
-    set.feature = function(feature) {
-      feature = private$sanitizeFeature(feature, self$predictor$data$feature.names)
-      private$flush()
-      private$setFeatureFromIndex(feature)
-      private$set.grid.size(private$grid.size.original)
-      self$run()
-    }
-  ), 
-  private = list(
-    aggregate = function() {
-      results = private$dataDesign[, self$feature.index, with = FALSE]
-      if (private$multiClass) {
-        y.hat.names = colnames(private$qResults)
-        results = cbind(results, private$qResults)
-        results = melt(results, variable.name = "..class.name", value.name = "y.hat", measure.vars = y.hat.names)
-      } else {
-        results[, "y.hat" := private$qResults]
-      }
-      if ("..class.name" %in% colnames(results)) {
-        results = results[, list(y.hat = mean(y.hat)), by = c(self$feature.name, "..class.name")]
-      } else {
-        results = results[, list(y.hat = mean(y.hat)), by = c(self$feature.name)]
-      }
-      results 
-    },
-    intervene = function() {
-      grid = get.1D.grid(private$dataSample[[self$feature.index[1]]], self$feature.type[1], self$grid.size[1])
-      
-      private$dataDesign.ids = rep(1:nrow(private$dataSample), times = length(grid))
-      dataDesign = private$dataSample[private$dataDesign.ids,]
-      dataDesign[, self$feature.index[1]] = rep(grid, each = nrow(private$dataSample))
-      
-      if (self$n.features == 2) {
-        grid2 = get.1D.grid(private$dataSample[[self$feature.index[2]]], self$feature.type[2], self$grid.size[2])
-        private$dataDesign.ids = rep(private$dataDesign.ids, times = length(grid2))
-        dataDesign2 = dataDesign[rep(1:nrow(dataDesign), times = length(grid2)), ]
-        dataDesign2[, self$feature.index[2]] = rep(grid2, each = nrow(dataDesign))
-        return(dataDesign2)
-      }
-      dataDesign
-    }, 
-    dataDesign.ids = NULL, 
-    grid.size.original = NULL,
-    setFeatureFromIndex = function(feature.index) {
-      self$feature.index = feature.index
-      self$n.features = length(feature.index)
-      self$feature.type = private$sampler$feature.types[feature.index]
-      self$feature.name = private$sampler$feature.names[feature.index]
-    },
-    printParameters = function() {
-      cat("features:", paste(sprintf("%s[%s]", self$feature.name, self$feature.type), collapse = ", "))
-      cat("\ngrid size:", paste(self$grid.size, collapse = "x"))
-    },
-    generatePlot = function() {
-      if (self$n.features == 1) {
-        p = ggplot(self$results, mapping = aes_string(x = self$feature.name,"y.hat")) + 
-          scale_y_continuous(expression(hat(y)))
-        if (self$feature.type == "numerical") {
-          p = p + geom_path() 
-        } else if (self$feature.type == "categorical") {
-          p = p + geom_point()
-        }
-        
-      } else if (self$n.features == 2) {
-        if (all(self$feature.type %in% "numerical") | all(self$feature.type %in% "categorical")) {
-          p = ggplot(self$results) + 
-            geom_tile(aes_string(x = self$feature.name[1], 
-              y = self$feature.name[2], 
-              fill = "y.hat")) + 
-            scale_fill_continuous(expression(hat(y)))
-        } else {
-          categorical.feature = self$feature.name[self$feature.type=="categorical"]
-          numerical.feature = setdiff(self$feature.name, categorical.feature)
-          p = ggplot(self$results) + 
-            geom_line(aes_string(x = numerical.feature, y = "y.hat", 
-              group = categorical.feature, color = categorical.feature)) + 
-            scale_y_continuous(expression(hat(y)))
-        }
-      }
-      if (private$multiClass) {
-        p = p + facet_wrap("..class.name")
-      } 
-      p
-    }, 
-    set.grid.size = function(size) {
-      self$grid.size = numeric(length=self$n.features)
-      names(self$grid.size) = private$sampler$feature.name[self$feature.index]
-      private$set.grid.size.single(size[1], 1)
-      if (self$n.features > 1) {
-        if (length(size) == 1) {
-          # If user only provided 1D grid size
-          private$set.grid.size.single(size[1], 2)
-        } else {
-          # If user provided 2D grid.size
-          private$set.grid.size.single(size[2], 2)
-        }
-      }
-    }, 
-    set.grid.size.single = function(size, feature.number) {
-      self$grid.size[feature.number] = ifelse(self$feature.type[feature.number] == "numerical", 
-        size, length(unique(private$sampler$get.x()[[self$feature.index[feature.number]]])))
-    }, 
-    sanitizeFeature = function(feature, feature.names) {
-      if (is.character(feature)) {
-        feature.char = feature
-        stopifnot(all(feature %in% feature.names))
-        feature = which(feature.char[1] == feature.names)
-        if (length(feature.char) == 2) {
-          feature = c(feature, which(feature.char[2] == feature.names))
-        }
-      }
-      feature
+      .Deprecated("Partial", msg = "The use of the 'PartialDependence' class is deprecated and it will 
+        be removed starting from version 0.4. Please use the 'Partial' class instead.")
+      super$initialize(predictor, feature, ice = FALSE, aggregation = "pdp", 
+        grid.size = grid.size, run = TRUE)
     }
   )
 )
@@ -258,6 +136,7 @@ PartialDependence = R6::R6Class("PartialDependence",
 #' plot.PartialDependence() plots the results of a PartialDependence object.
 #' 
 #' @param x A PartialDependence R6 object
+#' @param rug [logical] Should a rug be plotted to indicate the feature distribution?
 #' @return ggplot2 plot object
 #' @seealso 
 #' \link{PartialDependence}
@@ -274,8 +153,8 @@ PartialDependence = R6::R6Class("PartialDependence",
 #' # Plot the results directly
 #' plot(pdp.obj)
 #' }
-plot.PartialDependence = function(x) {
-  x$plot()
+plot.PartialDependence = function(x, rug = TRUE) {
+  x$plot(rug)
 }